{"schema":"decodex/commit/1","summary":"Extract scroll capture observation pipeline","authority":"manual"}
